\name{hsem-package}
\alias{hsem-package}
\docType{package}
\title{Hierarchical Structural Equation Models}
\description{
This package allows different models for multivariate reponse variables with a hierarchical structural equation models (HSEMs).
}

\details{
\tabular{ll}{
Package: \tab hsem\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-08-25\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
This is version 1.0 of the hsem package.
}


\author{Rezzy Eko Caraka, Maengseok Noh, Youngjo Lee

Maintainer: Rezzy Eko Caraka <rezzyekocaraka@gmail.com>
}


