\name{ml2}
\alias{ml2}
\docType{data}
\title{
simulated urge to smoke data 
}
\description{
By using an example for urge to smoke in the McNeish and
Hamaker (2020), this data set is a simulated subset. It
consists of 10 repetitions on regular time scales for 20
different individuals. For response variable, urge to Smoke is on a standardized scale average 0
and the standard deviation 1.}
\usage{data("ml2")}
\format{
  A data frame with 200 observations on the following 6 variables.
  \describe{
    \item{\code{urge}}{standardized urge to smoke} 
    \item{\code{urge1}}{previous urge to smoke} 
    \item{\code{dep}}{standardized depression} 
    \item{\code{dep1}}{ previous depression}
    \item{\code{id}}{ individual indicator}
    \item{\code{time}}{time indicator} 
  }
}


